% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam.R
\name{createExam}
\alias{createExam}
\title{createExam}
\usage{
createExam(filename, ASTAR = sample(10000, 1), MCMC = sample(10000, 1),
  FB = sample(10000, 1), VIT = sample(10000, 1), AB = sample(10000, 1),
  SCH = sample(10000, 1), MB = sample(10000, 1), ANN = sample(10000, 1),
  CNN = sample(10000, 1), TEXT = sample(10000, 1), NUMTEXT = 9,
  seedToExamRef = list(doRef = F))
}
\arguments{
\item{filename}{The name of the file to save the latex script in.}

\item{ASTAR}{The random seed for the A* question. Pass NA to not include the question.}

\item{MCMC}{The random seed for the Bayesian network/MCMC sampling question. Pass NA to not include the question.}

\item{FB}{The random seed for the HMM forward-backward algorithm question. Pass NA to not include the question.}

\item{VIT}{The random seed for the HMM Viterbi algorithm question. Pass NA to not include the question.}

\item{AB}{The random seed for the alpha-beta pruning question. Pass NA to not include the question. Including this
question will lead to the ab.jpg image file being created.}

\item{SCH}{The random seed for the scheduling question. Pass NA to not include the question.}

\item{MB}{The random seed for the multi-armed bandits question. Pass NA to not include the question.}

\item{ANN}{The random seed for the feed forward neural network question. Pass NA to not include the question. Including this
question will lead to the ffnn.jpg image file being created.}

\item{CNN}{The random seed for the CNN question. Pass NA to not include the question.}

\item{TEXT}{The random seed for selecting the text questions. Pass NA to not include any text questions.}

\item{NUMTEXT}{The number of text questions to include if the TEXT parameter is not NA. Should be 1 to 9.}

\item{seedToExamRef}{Used when generating the real exam. You should keep the default value when generating
practice exams.}
}
\description{
This function will generate the latex script and associated images for a complete exam. It will be used
to generate the real exam, and can be used by you to generate practice exams.
}
\details{
The real exam will contain 5-8 questions, selected so as to get a spread of questions from different
sections of the course and to not include pairs of question that are quite similar.

The set of questions include 9 algorithmic questions and 9 text questions. You can set seeds individually
for each algorithmic question, and these seeds determine the values generated for the question. If
you set NA the question is not included on the exam. You can specify the number of text questions and
the seed used to sample that number from the 9 available.

You will need to compile the latex script to generate a pdf. Make sure you keep the generated images in the
same folder as the tex file which compiling it. The images generated are the ab.jpg and ffnn.jpg. These
are generated for the alpha-beta pruning and feed forward neural networks questions respectively.

Note that this function sinks output streams to write to file. If it is cancelled during processing, you
may find your output is no longer directed to console. Use the sink function to redirect output streams back
to console.
}
