% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta.R
\name{makeQuestionAlphaBeta}
\alias{makeQuestionAlphaBeta}
\title{makeQuestionAlphaBeta}
\usage{
makeQuestionAlphaBeta(seed = NA, answer = T, latex = F, comments = T,
  imageFilename = NULL)
}
\arguments{
\item{seed}{The random seed to use to generate the values. If NA, they are generated from the
current seed.}

\item{answer}{Should the answer be outputed to console}

\item{latex}{Should latex script be outputed. This is for use with the makeExam function.}

\item{comments}{Should comments be outputed to console.}

\item{imageFilename}{The filename of the image to be saved when outputing latex script.
This is for use with the makeExam function.}
}
\description{
Make the alpha-beta pruning question
}
