% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStar.R
\name{makeQuestionAStar}
\alias{makeQuestionAStar}
\title{makeQuestionAStar}
\usage{
makeQuestionAStar(seed = NA, basic = F, nodes = 5, answer = T,
  latex = F)
}
\arguments{
\item{seed}{The random seed to use to generate the values. If NA, they are generated from the
current seed.}

\item{basic}{Should the heuristic be basic (assume all edges cost at least one). If false, a better heuristic
is sought.}

\item{nodes}{The number of nodes in the search space. The real question will have 5.}

\item{latex}{Should latex script be outputed. This is for use with the makeExam function.}
}
\description{
Make the A* question
}
