% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN.R
\name{makeQuestionFFNN}
\alias{makeQuestionFFNN}
\title{makeQuestionFFNN}
\usage{
makeQuestionFFNN(seed = NA, answer = T, latex = F, imageFilename = NULL)
}
\arguments{
\item{seed}{The random seed to use to generate the values. If NA, they are generated from the
current seed.}

\item{answer}{Should the answer be outputed to console}

\item{latex}{Should latex script be outputed. This is for use with the makeExam function.}

\item{imageFilename}{The filename of the image to be saved when outputing latex script.
This is for use with the makeExam function.}
}
\description{
Make the feed-forward neural network question
}
