% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiarmed_bandit.R
\name{makeQuestionMultiArmedBandit}
\alias{makeQuestionMultiArmedBandit}
\title{makeQuestionMultiArmedBandit}
\usage{
makeQuestionMultiArmedBandit(seed = NA, answer = T, latex = F)
}
\arguments{
\item{seed}{The random seed to use to generate the values. If NA, they are generated from the
current seed.}

\item{answer}{Should the answer be outputed to console}

\item{latex}{Should latex script be outputed. This is for use with the makeExam function.}
}
\description{
Make the multi-armed bandits question
}
