% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{makeQuestionSchedule}
\alias{makeQuestionSchedule}
\title{makeQuestionSchedule}
\usage{
makeQuestionSchedule(seed = NA, answer = T, latex = F)
}
\arguments{
\item{seed}{The random seed to use to generate the values. If NA, they are generated from the
current seed.}

\item{answer}{Should the answer be outputed to console}

\item{latex}{Should latex script be outputed. This is for use with the makeExam function.}
}
\description{
Make the scheduling question
}
