% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testQuestions.R
\name{makeTextQuestions}
\alias{makeTextQuestions}
\title{makeTextQuestions}
\usage{
makeTextQuestions(seed = NA, answer = T, latex = F, n = 1)
}
\arguments{
\item{seed}{The random seed to use to generate the values. If NA, they are generated from the
current seed.}

\item{answer}{Should the answer(s) be outputed to console}

\item{latex}{Should latex script be outputed. This is for use with the makeExam function.}

\item{n}{The number of questions to be selected}
}
\description{
Randomly select text questions from the available set.
}
