% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testQuestions.R
\name{makeTextQuestions2}
\alias{makeTextQuestions2}
\title{makeTextQuestions2}
\usage{
makeTextQuestions2(questions, answer = T)
}
\arguments{
\item{questions}{The questions to be selected. This should be an integer vector containing elements
between 1 and 9 inclusive.}

\item{answer}{Should the answer(s) be outputed to console}
}
\description{
Select specified text questions.
}
